backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
openai>=1.0.0
pymupdf>=1.24.0
youtube-transcript-api>=0.6.0
faiss-cpu>=1.8.0
numpy<2.0.0
python-dotenv>=1.0.0
tiktoken>=0.7.0
gtts>=2.5.0
requests>=2.31.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
easyocr>=1.7.0
pdf2image>=1.17.0
Pillow>=10.0.0
gunicorn==21.2.0
